crawler/requirements.txt
python-dotenv
supabase
fastapi
uvicorn
requests
aiohttp
beautifulsoup4
Pillow
google-api-python-client
